renamed resolve to get and improved some tests
