+++ lib/compile.js
@@ -1,4 +1,4 @@
-const { defaultResolver } = require('./resolve');
+const get = require('./get');
 
 const _defaultToString = Object.prototype.toString;
 /**
@@ -112,7 +112,7 @@ function inject(tokens, resolver) {
  * @returns {Compiler} - a function that accepts a view object and returns a
  *        rendered template string template
  */
-function compile(template, resolver = defaultResolver) {
+function compile(template, resolver = get) {
   // don't touch the template if it is not a string
   if (typeof template !== 'string') {
     return () => template;
@@ -121,7 +121,7 @@ function compile(template, resolver = defaultResolver) {
   // and resolver under the hood
   const tokens = tokenize(template)
   return function renderer(view) {
-    return inject(tokens, (varName) => resolver(varName, view))
+    return inject(tokens, (varName) => resolver(view, varName))
   }
 }
